Refactoring proto contract to use vitess

+++ backend/chess/board.go
@@ -170,10 +170,10 @@ func PbMoveStr(from string, to string) *pb.Move {
 	fromHex := HexStr(from)
 	toHex := HexStr(to)
 	return &pb.Move{
-		FromFile: int32(fromHex.File),
-		FromRank: int32(fromHex.Rank),
-		ToFile:   int32(toHex.File),
-		ToRank:   int32(toHex.Rank),
+		FromFile: fromHex.File,
+		FromRank: fromHex.Rank,
+		ToFile:   toHex.File,
+		ToRank:   toHex.Rank,
 	}
 }
